requirements.txt
# Core LangChain dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Google Generative AI (Gemini)
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Web Search Tool
tavily-python>=0.5.0

# PDF Generation
reportlab>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Email
# (built-in smtplib, email modules)

# Optional: Logging and utilities
pydantic>=2.0.0
typing-extensions>=4.0.0
